+++ doc/changelog.rst
@@ -18,6 +18,8 @@ Localization
 Admin
 ~~~~~
 
+- Add admin walkthrough
+
 Addons
 ~~~~~~